+++ Swiften/Elements/S5BProxyRequest.h
@@ -5,7 +5,7 @@
  */
 
 /*
- * Copyright (c) 2015-2016 Isode Limited.
+ * Copyright (c) 2015-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -30,7 +30,7 @@ public:
 public:
     struct StreamHost {
         std::string host;
-        int port;
+        unsigned short port;
         JID jid;
     };
