+++ Swiften/LinkLocal/DNSSD/Bonjour/BonjourQuerier.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2016 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -25,7 +25,7 @@ namespace Swift {
 
             std::shared_ptr<DNSSDBrowseQuery> createBrowseQuery();
             std::shared_ptr<DNSSDRegisterQuery> createRegisterQuery(
-                    const std::string& name, int port, const ByteArray& info);
+                    const std::string& name, unsigned short port, const ByteArray& info);
             std::shared_ptr<DNSSDResolveServiceQuery> createResolveServiceQuery(
                     const DNSSDServiceID&);
             std::shared_ptr<DNSSDResolveHostnameQuery> createResolveHostnameQuery(
