+++ Swiften/LinkLocal/LinkLocalService.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2016 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -31,7 +31,7 @@ namespace Swift {
                 return id.getName();
             }
 
-            int getPort() const {
+            unsigned short getPort() const {
                 return info.port;
             }
