+++ Swiften/LinkLocal/LinkLocalServiceBrowser.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2016 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -35,7 +35,7 @@ namespace Swift {
 
             void registerService(
                     const std::string& name,
-                    int port,
+                    unsigned short port,
                     const LinkLocalServiceInfo& info = LinkLocalServiceInfo());
             void updateService(
                     const LinkLocalServiceInfo& info = LinkLocalServiceInfo());
